tests/ocr-prototype/requirements.txt
# OCR Testing Backend Dependencies
# Smart Data Extractor (SME) Project

# Core Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0

# File handling and image processing
Werkzeug==3.0.1
Pillow==10.0.1
pdf2image==1.16.3

# OCR Services
pytesseract==0.3.10
google-cloud-vision==3.4.0
boto3==1.29.7

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Optional: For production deployment
gunicorn==21.2.0

# Development and testing (optional)
pytest==7.4.3
pytest-flask==1.3.0

# Required system dependencies (install separately):
# - tesseract-ocr
# - poppler-utils (for pdf2image)
#
# Installation commands:
# macOS: brew install tesseract poppler
# Ubuntu: sudo apt-get install tesseract-ocr poppler-utils
# Windows: Download and install from respective websites